refactor: extract camera app module
