+++ resources/js/app.js
@@ -1,7 +1,10 @@
 import './bootstrap';
 
 import Alpine from 'alpinejs';
+import cameraApp from './cameraApp';
 
 window.Alpine = Alpine;
+
+Alpine.data('cameraApp', cameraApp);
 
 Alpine.start();
